Refactor card, checkbox, input, radio, select stories
