+++ stories/components/card/card.sample.js
@@ -1,0 +1,16 @@
+export default `
+  <div class="ch-card">
+    <header class="ch-card-header">
+      <h4>Lorem ipsum dolor sit amet</h4>
+    </header>
+
+    <main class="ch-card-content">
+      <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Dolor, maiores.</p>
+    </main>
+
+    <footer class="ch-card-footer">
+      <button type="button" class="ch-button ch-button--primary">Primary</button>
+      <button type="button" class="ch-button ch-button--secondary">Secondary</button>
+    </footer>
+  </div>
+`
