+++ stories/components/input/input.sample.js
@@ -1,0 +1,11 @@
+const html =  `<input type="text" placeholder="Insert data" class="ch-input">`
+
+const htmlWithIcons = `
+  <span class="ch-input has-leadingIcon has-trailingIcon">
+    <i class="ch-icon fal fa-search"></i>
+    <input type="text" placeholder="Insert data" />
+    <i class="ch-spinner"></i>
+  </span>
+`
+
+export { html, htmlWithIcons }
