+++ stories/components/radio/radio.stories.js
@@ -1,0 +1,14 @@
+import { withKnobs } from '@storybook/addon-knobs'
+import html from './radio.sample'
+
+export default {
+  title: 'Components/Radio',
+  decorators: [withKnobs],
+  parameters: {
+    docs: {
+      page: ''
+    }
+  }
+}
+
+export const Radio = () => html
