+++ stories/components/select/select.sample.js
@@ -1,0 +1,7 @@
+export default `
+  <select class="ch-select">
+    <option value="1">Option 1</option>
+    <option value="2" selected="">Option 2</option>
+    <option value="3">Option 3</option>
+  </select>
+`
